Rearchitecture of Cell to align more to single-responsibility
Splits cell into a cellnetwork and a set of rules. Creates a RuleEngine
to operate on rules which are now objects in and of themselves. History
now stores this rule object and it can be replayed directly out of the
history. Speculative solution becomes a "rule", such that it can be
storeed in the history and replayed without any special workarounds.